get async worker to work with autofix e2e
