+++ src/seer/tasks.py
@@ -17,11 +17,14 @@ from sentry_sdk.integrations.asyncio import AsyncioIntegration
 from sqlalchemy import func, select
 from sqlalchemy.ext.asyncio import async_sessionmaker
 
+import seer.automation.autofix.tasks
 from seer.db import AsyncSession, ProcessRequest
+from seer.task_factory import AsyncTaskFactory, _async_task_factories
 
 _A = TypeVar("_A")
 
 logger = logging.getLogger("asyncrunner")
+logger.addHandler(logging.StreamHandler())
 logger.setLevel(logging.INFO)
 
 
@@ -57,86 +60,16 @@ class Period:
         return Period(period_start=self.period_start - self.period, period=self.period)
 
 
-class AsyncTaskFactory(abc.ABC):
-    @abc.abstractmethod
-    def matches(self, process_request: ProcessRequest) -> bool:
-        pass
-
-    @abc.abstractmethod
-    async def invoke(self, process_request: ProcessRequest):
-        pass
-
-    async def async_celery_job(self, cb: Callable[[], celery.result.AsyncResult]):
-        logger.info("Starting async celery job")
-        loop = asyncio.get_running_loop()
-        q: Queue = Queue(1)
-
-        with ThreadPoolExecutor() as pool:
-            ar = await loop.run_in_executor(pool, cb)
-
-            def run():
-                def on_message(raw: Any):
-                    # Keep trying to put the item into the queue by removing existing item until we get in.
-                    logger.info("Received response from celery job")
-                    while True:
-                        try:
-                            q.put_nowait(raw)
-                            return
-                        except Full:
-                            try:
-                                q.get_nowait()
-                            except Empty:
-                                pass
-
-                ar.get(on_message=on_message, propagate=True)
-
-            complete = loop.run_in_executor(pool, run)
-
-            try:
-                while not complete.done():
-                    get = loop.run_in_executor(pool, q.get)
-                    await asyncio.wait([get, complete], return_when=asyncio.FIRST_COMPLETED)
-                    if get.done():
-                        v = await get
-                        if v:
-                            if v["status"] == "PROGRESS":
-                                try:
-                                    yield v["result"]
-                                except Exception as e:
-                                    if not complete.done():
-                                        logger.warning("SIGUSR1 on job, generator failed")
-                                        await loop.run_in_executor(
-                                            pool,
-                                            lambda: ar.revoke(
-                                                terminate=True, signal="SIGUSR1", wait=False
-                                            ),
-                                        )
-                                    raise e
-                            # if v['status'] == 'FAILURE':
-                            #     raise v['result']
-            finally:
-                logger.info("async celery job completing")
-                loop.run_in_executor(pool, lambda: q.put_nowait(None))
-        await complete
-
-
-_async_task_factories: list[Callable[[], AsyncTaskFactory]] = []
-
-
-def async_task_factory(f: Callable[[], AsyncTaskFactory]) -> Callable[[], AsyncTaskFactory]:
-    _async_task_factories.append(f)
-    return f
-
-
 @dataclasses.dataclass
 class AsyncApp:
     end_event: asyncio.Event = dataclasses.field(default_factory=lambda: asyncio.Event())
     num_consumers: int = 10
     queue: asyncio.Queue = dataclasses.field(default_factory=lambda: asyncio.Queue())
-    task_factories: list[Callable[[], AsyncTaskFactory]] = dataclasses.field(
-        default_factory=lambda: _async_task_factories,
-    )
     consumer_sleep: int = 5
+
+    @property
+    def task_factories(self) -> list[Callable[[], AsyncTaskFactory]]:
+        return _async_task_factories
 
     async def run_or_end(self, c: Future[_A] | Coroutine[Any, Any, _A]) -> tuple[_A] | None:
         end_task = asyncio.create_task(self.end_event.wait())
@@ -168,7 +101,8 @@ class AsyncApp:
                         )
                     )
                 )
-            if result is not None:
+            if result is not None and result[0]:
+                logger.info(f"Found {len(result[0])} process requests")
                 for item in result[0]:
                     logger.info(f"Picked up process request, running")
                     await self.run_or_end(self.queue.put(item))
@@ -198,6 +132,7 @@ class AsyncApp:
                 continue
 
             item: ProcessRequest = result[0]
+            logger.info(f"Consumer loop working on {item.name}.")
             for factory in self.task_factories:
                 task = factory()
                 accept = task.matches(item)
