+++ supervisord.conf
@@ -25,18 +25,6 @@ stderr_logfile=/dev/stderr
 stderr_logfile_maxbytes=0
 
 [program:asyncworker]
-command=/app/celeryworker.sh
-directory=/app
-startsecs=0
-autostart=true
-autorestart=unexpected
-exitcodes=0
-stdout_logfile=/dev/stdout
-stdout_logfile_maxbytes=0
-stderr_logfile=/dev/stderr
-stderr_logfile_maxbytes=0
-
-[program:asyncworker]
 command=/app/asyncworker.sh
 directory=/app
 startsecs=0
